requirements.txt
asgiref==3.4.1
Django==3.2.7
inertia-django==0.3.2
python-dotenv==0.19.0
pytz==2021.1
sqlparse==0.4.2
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.6
Django==3.2.7
django-js-routes==0.1.4
django-meta==2.0.0
idna==3.2
inertia-django==0.3.2
mysqlclient==2.0.3
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.6
Django==3.2.7
django-debug-toolbar==3.2.2
django-js-routes==0.1.4
django-meta==2.0.0
djangomix==1.1.3
djangorestframework==3.12.4
idna==3.2
inertia-django==0.3.2
mysqlclient==2.0.3
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
amqp==5.0.6
asgiref==3.4.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
billiard==3.6.4.0
celery==5.1.2
certifi==2021.5.30
cffi==1.14.6
channels==3.0.4
charset-normalizer==2.0.6
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
contextlib2==21.6.0
cryptography==3.4.8
daphne==3.0.2
Django==3.2.7
django-debug-toolbar==3.2.2
django-js-routes==0.1.4
django-meta==2.0.0
djangomix==1.1.3
djangorestframework==3.12.4
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
inertia-django==0.3.2
kombu==5.1.0
marshmallow==3.13.0
mysqlclient==2.0.3
ndg-httpsclient==0.5.1
Pillow==8.3.2
prompt-toolkit==3.0.20
pusher==3.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyNaCl==1.4.0
pyOpenSSL==20.0.1
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
schema==0.7.4
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
swapper==1.1.2.post1
Twisted==21.7.0
txaio==21.2.1
typing-extensions==3.10.0.2
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
zope.interface==5.4.0
